Remove recently introduced extra copy/move. Fixes move-semantics tests
